+++ src/main/java/net/dv8tion/jda/internal/entities/channel/AbstractStandardGuildMessageChannelImpl.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package net.dv8tion.jda.internal.entities;
+package net.dv8tion.jda.internal.entities.channel;
 
-import net.dv8tion.jda.internal.entities.mixin.channel.middleman.StandardGuildMessageChannelMixin;
+import net.dv8tion.jda.internal.entities.GuildImpl;
+import net.dv8tion.jda.internal.entities.channel.mixin.middleman.StandardGuildMessageChannelMixin;
 
 import javax.annotation.Nullable;
