+++ src/main/java/net/dv8tion/jda/internal/entities/channel/mixin/middleman/StandardGuildChannelMixin.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package net.dv8tion.jda.internal.entities.mixin.channel.middleman;
+package net.dv8tion.jda.internal.entities.channel.mixin.middleman;
 
 import net.dv8tion.jda.api.entities.channel.middleman.StandardGuildChannel;
-import net.dv8tion.jda.internal.entities.mixin.channel.attribute.ICategorizableChannelMixin;
-import net.dv8tion.jda.internal.entities.mixin.channel.attribute.IInviteContainerMixin;
-import net.dv8tion.jda.internal.entities.mixin.channel.attribute.IPermissionContainerMixin;
-import net.dv8tion.jda.internal.entities.mixin.channel.attribute.IPositionableChannelMixin;
+import net.dv8tion.jda.internal.entities.channel.mixin.attribute.ICategorizableChannelMixin;
+import net.dv8tion.jda.internal.entities.channel.mixin.attribute.IInviteContainerMixin;
+import net.dv8tion.jda.internal.entities.channel.mixin.attribute.IPermissionContainerMixin;
+import net.dv8tion.jda.internal.entities.channel.mixin.attribute.IPositionableChannelMixin;
 
 public interface StandardGuildChannelMixin<T extends StandardGuildChannelMixin<T>> extends
         StandardGuildChannel,
