+++ src/main/java/net/dv8tion/jda/internal/handle/StageInstanceDeleteHandler.java
@@ -21,7 +21,7 @@ import net.dv8tion.jda.api.events.stage.StageInstanceDeleteEvent;
 import net.dv8tion.jda.api.utils.data.DataObject;
 import net.dv8tion.jda.internal.JDAImpl;
 import net.dv8tion.jda.internal.entities.GuildImpl;
-import net.dv8tion.jda.internal.entities.StageChannelImpl;
+import net.dv8tion.jda.internal.entities.channel.concrete.StageChannelImpl;
 
 public class StageInstanceDeleteHandler extends SocketHandler
 {
